% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_function.R
\name{product_function}
\alias{product_function}
\title{Return the product of two numbers}
\usage{
product_function(a, b)
}
\arguments{
\item{a}{Integer or numeric. First value to multiply.}

\item{b}{Integer or numeric. Second value to multiply.}
}
\value{
An integer or numeric. The product of a and b
}
\description{
Takes to numbers (numeric or integer) and returns
the product
}
\examples{
#Input integers
product_function(1L, 2L)

#Input numeric
product_function(1.0, 2.5)
}
