% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_function.R
\name{product_function}
\alias{product_function}
\title{Return the product of two numbers}
\usage{
product_function(a, b)
}
\arguments{
\item{a}{}

\item{b}{}
}
\value{

}
\description{
Return the product of two numbers
}
